NumberFromString now in util
